+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,15 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.lumebits.corremundos">
+
+    <uses-permission android:name="android.permission.INTERNET" />
+    <queries>
+        <!-- If your app opens https URLs -->
+        <intent>
+            <action android:name="android.intent.action.VIEW" />
+            <data android:scheme="https" />
+        </intent>
+    </queries>
+
    <application
         android:label="${appName}"
         android:icon="@mipmap/launcher_icon">
